re-arrange the code a bit
